Streamline usages of spki:: and der:: types which we override anyways
