+++ src/types/der/asn1/uint.rs
@@ -1,7 +1,3 @@
 // This Source Code Form is subject to the terms of the Mozilla Public
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at https://mozilla.org/MPL/2.0/.
-
-pub mod ia5string;
-
-pub use ia5string::*;
