+++ src/types/der/mod.rs
@@ -2,5 +2,4 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at https://mozilla.org/MPL/2.0/.
 
-pub mod der;
-pub mod spki;
+pub mod asn1;
